evaluation/MedEvalKit_local/models/MedPLIB/requirements.txt

absl-py==2.1.0
accelerate==0.26.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cupy-cuda12x==12.1.0
cycler==0.12.1
deepspeed==0.13.1
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
fastapi==0.110.0
fastrlock==0.8.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.62.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib_resources==6.4.0
interegular==0.3.3
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
lazy_loader==0.3
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
#mpi4py==3.1.4
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
openai==1.14.2
opencv-python==4.9.0.80
orjson==3.9.15
outlines==0.0.36
packaging==24.0
pandas==2.2.1
peft==0.10.0
pillow==10.2.0
plumbum==1.8.2
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools==2.0.7
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
ray==2.10.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rpyc==6.0.0
ruff==0.3.4
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.43.0
setproctitle==1.3.3
sglang==0.1.13
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.6.13
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
transformers==4.31.0
triton==2.1.0
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
#vllm==0.3.3
wandb==0.16.4
watchfiles==0.21.0
#wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.0.1
#xformers==0.0.23
yarl==1.9.4
zmq==0.0.0
